Cleaned up warnings

+++ hlsl-shader-tracking/src/opvariable_init.cpp
@@ -160,13 +160,13 @@ void init_tiny_renderer(GLFWwindow* window)
     auto vert = load_file(k_asset_dir + "opvariable_init.vs.spv");
     auto frag = load_file(k_asset_dir + "opvariable_init.ps.spv");
     tr_create_shader_program(m_renderer, 
-                             vert.size(), (uint32_t*)(vert.data()), "vsmain", 
-                             frag.size(), (uint32_t*)(frag.data()), "psmain", &m_shader);
+                             (uint32_t)vert.size(), (uint32_t*)(vert.data()), "vsmain", 
+                             (uint32_t)frag.size(), (uint32_t*)(frag.data()), "psmain", &m_shader);
 #elif defined(TINY_RENDERER_DX)
     auto hlsl = load_file(k_asset_dir + "opvariable_init.hlsl");
     tr_create_shader_program(m_renderer, 
-                             hlsl.size(), hlsl.data(), "vsmain", 
-                             hlsl.size(), hlsl.data(), "psmain", &m_shader);
+                             (uint32_t)hlsl.size(), hlsl.data(), "vsmain", 
+                             (uint32_t)hlsl.size(), hlsl.data(), "psmain", &m_shader);
 #endif
 
     tr_vertex_layout vertex_layout = {};
@@ -191,8 +191,8 @@ void init_tiny_renderer(GLFWwindow* window)
         vertexData[4*1 + 0] += -0.5f;
         vertexData[4*2 + 0] += -0.5f;
 
-        uint64_t vertexDataSize = sizeof(float) * vertexData.size();
-        uint64_t vertexStride = sizeof(float) * 4;
+        uint32_t vertexDataSize = sizeof(float) * (uint32_t)vertexData.size();
+        uint32_t vertexStride = sizeof(float) * 4;
         tr_create_vertex_buffer(m_renderer, vertexDataSize, true, vertexStride, &m_tri_vertex_buffer);
         memcpy(m_tri_vertex_buffer->cpu_mapped_address, vertexData.data(), vertexDataSize);
     }
@@ -211,8 +211,8 @@ void init_tiny_renderer(GLFWwindow* window)
         vertexData[4*2 + 0] += 0.5f;
         vertexData[4*3 + 0] += 0.5f;
 
-        uint64_t vertexDataSize = sizeof(float) * vertexData.size();
-        uint64_t vertexStride = sizeof(float) * 4;
+        uint32_t vertexDataSize = sizeof(float) * (uint32_t)vertexData.size();
+        uint32_t vertexStride = sizeof(float) * 4;
         tr_create_vertex_buffer(m_renderer, vertexDataSize, true, vertexStride, &m_rect_vertex_buffer);
         memcpy(m_rect_vertex_buffer->cpu_mapped_address, vertexData.data(), vertexDataSize);
         
@@ -221,7 +221,7 @@ void init_tiny_renderer(GLFWwindow* window)
             0, 2, 3
         };
         
-        uint64_t indexDataSize = sizeof(uint16_t) * indexData.size();
+        uint32_t indexDataSize = sizeof(uint16_t) * (uint32_t)indexData.size();
         tr_create_index_buffer(m_renderer, indexDataSize, true, tr_index_type_uint16, &m_rect_index_buffer);
         memcpy(m_rect_index_buffer->cpu_mapped_address, indexData.data(), indexDataSize);
     }
@@ -249,7 +249,7 @@ void draw_frame()
 
     tr_begin_cmd(cmd);
     tr_cmd_render_target_transition(cmd, render_target, tr_texture_usage_present, tr_texture_usage_color_attachment); 
-    tr_cmd_set_viewport(cmd, 0, 0, s_window_width, s_window_height, 0.0f, 1.0f);
+    tr_cmd_set_viewport(cmd, 0, 0, (float)s_window_width, (float)s_window_height, 0.0f, 1.0f);
     tr_cmd_set_scissor(cmd, 0, 0, s_window_width, s_window_height);
     tr_cmd_begin_render(cmd, render_target);
     tr_clear_value clear_value = {0.0f, 0.0f, 0.0f, 0.0f};
